Externally-sourced JSON files
New driver program processes JSON files in a ../Out/external folder and
formats the results. This allows for use in conjunction with testing
programs in any language. Also added new functionality to specify
expected value as UNTESTED, meaning show the actual value but do not
assert for that field (useful if automated testing difficult for
specific fields). Four examples taken from an Oracle project are
included.